refactor: Introduce barrel files to simplify module imports
