+++ src/components/common/index.ts
@@ -1,0 +1,6 @@
+export { default as Button } from './Button';
+export { default as Input } from './Input';
+export { default as Logo } from './Logo';
+export { default as SearchInput } from './SearchInput';
+
+export * from './link';
