+++ src/components/common/link/index.ts
@@ -1,0 +1,2 @@
+export { default as ExternalLink } from './ExternalLink';
+export { default as InternalLink } from './InternalLink';
