+++ src/components/layout/Layout.tsx
@@ -1,5 +1,4 @@
-import Header from '@/components/layout/header/Header';
-import Footer from '@/components/layout/footer/Footer';
+import { Header, Footer } from '@/components/layout';
 import { Outlet } from 'react-router-dom';
 
 export default function Layout() {
