+++ src/components/layout/banner/index.ts
@@ -1,0 +1,1 @@
+export { default as HeroBanner } from './HeroBanner';
