+++ src/components/layout/footer/FooterContent.tsx
@@ -1,6 +1,4 @@
-import FooterAssistance from '@/components/layout/footer/FooterAssistance';
-import FooterLogo from '@/components/layout/footer/FooterLogo';
-import FooterServices from '@/components/layout/footer/FooterServices';
+import { FooterLogo, FooterServices, FooterAssistance } from '@/components/layout/footer';
 
 interface FooterContentProps {
   className?: string;
