+++ src/components/layout/footer/index.ts
@@ -1,0 +1,6 @@
+export { default as Footer } from './Footer';
+export { default as FooterAssistance } from './FooterAssistance';
+export { default as FooterContent } from './FooterContent';
+export { default as FooterLogo } from './FooterLogo';
+export { default as FooterServices } from './FooterServices';
+export { default as SocialLinks } from './SocialLinks';
