+++ src/components/layout/header/NavMain.tsx
@@ -1,4 +1,4 @@
-import { Link } from 'react-router-dom';
+import { InternalLink } from '@/components/common';
 
 interface NavMainProps {
   className?: string;
@@ -7,18 +7,18 @@ interface NavMainProps {
 export default function NavMain({ className }: NavMainProps) {
   return (
     <div className={`flex flex-row gap-10 ${className ?? ''}`}>
-      <Link to='/' className='text-black'>
+      <InternalLink to='/' className='text-black'>
         Home
-      </Link>
-      <Link to='/about' className='text-gray-400'>
+      </InternalLink>
+      <InternalLink to='/about' className='text-gray-400'>
         About
-      </Link>
-      <Link to='/contact' className='text-gray-400'>
+      </InternalLink>
+      <InternalLink to='/contact' className='text-gray-400'>
         Contact Us
-      </Link>
-      <Link to='/blog' className='text-gray-400'>
+      </InternalLink>
+      <InternalLink to='/blog' className='text-gray-400'>
         Blog
-      </Link>
+      </InternalLink>
     </div>
   );
 }
