+++ src/components/layout/header/index.ts
@@ -1,0 +1,4 @@
+export { default as Header } from './Header';
+export { default as NavMain } from './NavMain';
+export { default as NavSub } from './NavSub';
+export { default as UserActions } from './UserActions';
