+++ src/components/layout/index.ts
@@ -1,0 +1,5 @@
+export { default as Layout } from './Layout';
+
+export * from './banner';
+export * from './header';
+export * from './footer';
