+++ src/pages/index.ts
@@ -1,0 +1,13 @@
+export { default as AboutPage } from './AboutPage';
+export { default as BlogPage } from './BlogPage';
+export { default as CamerasPage } from './CamerasPage';
+export { default as CartPage } from './CartPage';
+export { default as ComputersPage } from './ComputersPage';
+export { default as ContactPage } from './ContactPage';
+export { default as GamingPage } from './GamingPage';
+export { default as HeadphonesPage } from './HeadphonesPage';
+export { default as HomePage } from './HomePage';
+export { default as LikedPage } from './LikedPage';
+export { default as PhonesPage } from './PhonesPage';
+export { default as ProfilePage } from './ProfilePage';
+export { default as SmartWatchesPage } from './SmartWatchesPage';
